requirements.txt
asyncpraw==7.8.1
python-dotenv==1.0.0
spotipy==2.22.1
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
black>=23.0.0
pre-commit>=3.5.0